--- a/auto_gptq/nn_modules/qlinear_old.py
+++ b/auto_gptq/nn_modules/qlinear_old.py
@@ -49,7 +49,9 @@ class QuantLinear(nn.Module):
             
         self.kernel_switch_threshold = kernel_switch_threshold
         self.is_cuda = is_cuda
-
+        if infeatures % 256 != 0 or outfeatures % 256 != 0:
+            self.quant_cuda_available = False
+            
     def pack(self, linear, scales, zeros, g_idx):
         scales = scales.t().contiguous()
         zeros = zeros.t().contiguous()