Provides faster speed than triton if act-order and groupsize are not used at the same time.