@PanQiWei 2023-05-01
*And sorry for the long delay response for I currently on a vocation with my families.*
Excellent! This big improvement can really help people who can't access to triton to trade-off between speed and quantized model's quality. I'll look into the code details and merge this great work to main branch as soon as possible. 🥂