@TheBloke 2023-05-01
I am not seeing speed improvements with this code at all.  But it does fix CUDA inference which was previously not working reliably for me through text-generation-webui (via @Laaza's PR https://github.com/oobabooga/text-generation-webui/pull/1668)

# Summary
* With base AutoGPTQ + CUDA + no-act-order model = 
  * Gibberish output
  * Slow speed
* With this PR + CUDA + same model =
  * Good output
  * Slow speed
* With base AutoGPTQ + Triton + same model =
  * Good output
  * Better speed     

## Testing setup

* NVidia 4090 24GB in Ubuntu
* This model: [TheBloke/wizardLM-7B-GPTQ/wizardLM-7B-GPTQ-4bit.compat.no-act-order.safetensors](https://huggingface.co/TheBloke/wizardLM-7B-GPTQ/blob/main/wizardLM-7B-GPTQ-4bit.compat.no-act-order.safetensors)
  * Bits = 4
  * Groupsize = 128
  * No act-order/desc_act 

## How I tested:
```
root@ad62753e041d:~/AutoGPTQ# pip uninstall auto-gptq
Found existing installation: auto-gptq 0.1.0.dev0
Uninstalling auto-gptq-0.1.0.dev0:
  Would remove:
    /usr/local/lib/python3.10/dist-packages/auto_gptq-0.1.0.dev0.dist-info/*
    /usr/local/lib/python3.10/dist-packages/auto_gptq/*
    /usr/local/lib/python3.10/dist-packages/quant_cuda.cpython-310-x86_64-linux-gnu.so
Proceed (Y/n)? Y
  Successfully uninstalled auto-gptq-0.1.0.dev0
WARNING: Running pip as the 'root' user can result in broken permissions and conflicting behaviour with the system package manager. It is recommended to use a virtual environment instead: https://pip.pypa.io/warnings/venv

root@ad62753e041d:~/AutoGPTQ# cd ../auto-new-cuda/
root@ad62753e041d:~/auto-new-cuda# git branch
* faster-cuda-no-actorder
  main
root@ad62753e041d:~/auto-new-cuda# pip install .
....
Successfully installed auto-gptq-0.1.0.dev0
```
## Executing text-generation-webui like this:

```
root@ad62753e041d:~/textgen-auto# python server.py --model wiz-no-act  --autogptq  --listen --wbits 4 --groupsize 128 --model_type llama
Gradio HTTP request redirected to localhost :)
bin /usr/local/lib/python3.10/dist-packages/bitsandbytes/libbitsandbytes_cuda116.so
Loading wiz-no-act...
Quantize config found for wiz-no-act. Setting wbits=4 and groupsize=128.
Loading quantized model with AutoGPTQ from models/wiz-no-act/wizardLM-7B-GPTQ-4bit.no-act-order.safetensors
The safetensors archive passed at models/wiz-no-act/wizardLM-7B-GPTQ-4bit.no-act-order.safetensors does not contain metadata. Make sure to save your model with the `save_pretrained` method. Defaulting to 'pt' metadata.
Loaded the model in 2.86 seconds.
Running on local URL:  http://0.0.0.0:7860
```

## Results

### With base AutoGPTQ

commit = `144bd804369264998dce2ca5b97a365d01a725a6`

#### Speed = 7-8 t/s
```
Output generated in 16.75 seconds (7.82 tokens/s, 131 tokens, context 16, seed 1472144110)
```

#### Output = gibberish / no good
```
### Human: write a story about llamas
### Assistant:As the weather was pretty, as an eye-viewing ship, and they’re recommendations for the day. The Sailor here's that were just one of those things, like the one with the name “Andy” (their good luck) or "No" in the snow.
The 2019 was my own unique(ish) opportunity (U) and the best?
As an AI field, the user is able to see if we _are_dressed*? The 2019 llama was my own unique(ish) opportunity (U) and the best?
```

### With this PR #38

commit = `f0f37c1fe7fb17a8195ac975e1129c33082321ba`

#### Speed = 6-7 t/s  SLOW
```
Output generated in 37.39 seconds (6.55 tokens/s, 245 tokens, context 16, seed 1927118888)
Output generated in 78.88 seconds (6.28 tokens/s, 495 tokens, context 16, seed 648098553)
```
#### Output = GOOD
```
### Human: write a story about llamas
### Assistant:Once upon a time, in a faraway land, there lived a herd of llamas. They roamed the fields and hillsides all day long, grazing on fresh grass and sleeping under the stars at night. One day, a wise old llama named Luna noticed that some of the younger llamas were getting bored with their daily routine. "There must be more to life than this," they complained. Luna knew just what to do. She called together all the llamas and said, "Let us go on an adventure!" The other llamas were hesitant at first, but Luna promised them it would be fun. So off they set, across the mountains and through the forests until they came to a clearing where they saw something they had never seen before - a human! The humans were amazed to see such tall creatures with spots on their coats. The llamas were equally surprised by the strange animals with two legs instead of four. Together, they explored the world beyond their own and discovered new wonders around every corner. When they returned home, the llamas told everyone about their incredible journey. And from that day forward, they knew that there was always something exciting waiting just over the next hill.
```

### With Triton

#### Executed using base AutoGPTQ Triton like this
```
root@ad62753e041d:~/AutoGPTQ# pip uninstall -qy auto-gptq && pip install -q .

root@ad62753e041d:~/AutoGPTQ# git log | head -1
commit 144bd804369264998dce2ca5b97a365d01a725a6

root@ad62753e041d:~/AutoGPTQ# cd ~/textgen-auto
root@ad62753e041d:~/textgen-auto# python server.py --model wiz-no-act  --autogptq --autogptq-triton  --listen --wbits 4 --groupsize 128 --model_type llama
Gradio HTTP request redirected to localhost :)
bin /usr/local/lib/python3.10/dist-packages/bitsandbytes/libbitsandbytes_cuda116.so
Loading wiz-no-act...
Quantize config found for wiz-no-act. Setting wbits=4 and groupsize=128.
Loading quantized model with AutoGPTQ from models/wiz-no-act/wizardLM-7B-GPTQ-4bit.no-act-order.safetensors
use_triton will force moving the whole model to GPU, make sure you have enough VRAM.
The safetensors archive passed at models/wiz-no-act/wizardLM-7B-GPTQ-4bit.no-act-order.safetensors does not contain metadata. Make sure to save your model with the `save_pretrained` method. Defaulting to 'pt' metadata.
100%|█████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 12/12 [00:39<00:00,  3.28s/it]
Loaded the model in 42.15 seconds.
Running on local URL:  http://0.0.0.0:7860
```

#### Speed = 11-12t/s  BETTER (but slower than GPTQ-for-LLaMa)
```
Output generated in 28.00 seconds (11.46 tokens/s, 321 tokens, context 16, seed 476875609)
Output generated in 30.73 seconds (12.27 tokens/s, 377 tokens, context 16, seed 1175561351)
```

#### Output = GOOD
```
### Human: write a story about llamas
### Assistant:Once upon a time, in a faraway land, there lived a herd of llamas. They roamed the vast plains and mountains with their long necks swinging from side to side. The llamas were known for being gentle creatures who loved to graze on grass all day. One sunny morning, as they were out grazing, they noticed something strange in the distance. It was a group of humans carrying tools and building materials. 
The llamas were curious and approached the humans cautiously. To their surprise, the humans spoke to them in a language they had never heard before. The llamas were confused but listened intently as the humans explained that they were building a wall around their village. The llamas were puzzled but didn't understand why they needed a wall. 
The humans explained that they were protecting themselves from invaders who wanted to steal their land and resources. The llamas realized how important it was to be protected and offered to help the humans build the wall. Over the next few days, the llamas worked tirelessly alongside the humans, using their strong legs to carry rocks and dirt. 
As they finished the wall, the humans thanked the llamas and gave them some food as a reward. The llamas were happy to have helped and continued to graze peacefully on the plains, feeling proud of themselves for contributing to the safety of the human community. From that day forward, the llamas became known as guardians of the land, protecting it from harm.
```